Move bundle functions to bundler.ts
